backend/requirements.txt
requests>=2.31.0
asyncio-mqtt>=0.11.0
gremlinpython>=3.6.5
chromadb>=0.4.15
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.33.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
statistics>=1.0.3.5
json5>=0.9.14
python-dateutil>=2.8.2
typing-extensions>=4.7.0
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0
Authlib>=1.2.1
python-dotenv>=1.0.0
azure-cosmos>=4.5.0
azure-storage-blob>=12.17.0
azure-core>=1.29.0
ipython>=8.14.0
jupyter>=1.0.0
loguru>=0.7.0
fastapi>=0.103.0
uvicorn>=0.23.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
PyPDF2>=3.0.0